SDL with SDL
