+++ source/utils.hpp
@@ -1,3 +1,15 @@
+
+bool isFileExist(const char *file)
+{
+	struct stat	st = { 0 };
+
+	if (stat(file, &st) == -1) {
+		return (false);
+	}
+
+	return (true);
+}
+
 void close()
 {
 	//Free loaded images
